requirements.txt
tensorflow==1.13.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
numpy==1.19.4
h5py==2.10.0
opencv-python
Pillow
flask
matplotlib
torch
torchvision
